pyproject.toml
[project]
name = "square-cli"
version = "0.1.0"
description = "A CLI for Square — manage catalog, sales, inventory, customers, and more from the command line."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Hugh Gramelspacher" },
]
keywords = ["square", "pos", "cli", "catalog", "payments", "inventory"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
]
dependencies = [
    "squareup>=40.0",
    "typer>=0.12",
    "rich>=13.0",
    "keyring>=25.0",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
square = "square_cli.cli:app"

[project.urls]
Homepage = "https://github.com/hughgramel/square-cli"
Repository = "https://github.com/hughgramel/square-cli"
Issues = "https://github.com/hughgramel/square-cli/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/square_cli"]

[tool.ruff]
target-version = "py311"
line-length = 100
